--- src/types/Post.ts
@@ -1,8 +1,0 @@
-export interface Post {
-  id: string;
-  name: string;
-  message: string;
-  email?: string;
-  createdAt?: string;
-  updatedAt?: string;
-}
